+++ django_typer/types.py
@@ -17,7 +17,7 @@ from django.core.management import CommandError
 from django.utils.translation import gettext_lazy as _
 from typer import Option
 
-from .completers import complete_import_path
+from .completers import complete_directory, complete_import_path
 
 COMMON_PANEL = "Django"
 
@@ -137,6 +137,7 @@ PythonPath = Annotated[
             ),
         ),
         rich_help_panel=COMMON_PANEL,
+        shell_complete=complete_directory,
     ),
 ]
 """